refactor: Move AnnotationsSlice logic from hook to util
